Add Windows path manipulation (#4)
This makes both Windows and Posix path modules available as `winpath` and `posixpath`, respectively. The appropriate implementation for the `path` module is chosen at compile time. This makes it easy to test both Posix path manipulation and Windows path manipulation on any platform. Much of the path manipulation code is shared between the two, making the Posix- and Windows-specific bits very small.