Normalize paths everywhere
While this is not mandatory for correctness, it cleans up the output of
the rules.

This change also changes the behavior of `:path()`. It will no longer
return a path with a leading dot (e.g., './foo'). If a built executable
needs to be executed in a task, the path needs to be joined with the
dot (e.g., `path.join('.', 'foo')`).